requirements.txt
# prompt-linter/training/requirements.txt

# Core ML Framework
torch

# Hugging Face library for transformer models and tokenizers
transformers

# Library for tokenizing text efficiently
tokenizers

# Library for Explainable AI (XAI)
captum

# Data manipulation and analysis
pandas

# For writing the final rules.yaml file
PyYAML

# Progress bar library for loops
tqdm

# Web Server
fastapi
uvicorn[standard]

# Additional Libraries
datasets
pyarrow
pandas
boto3